+++ database/migrations/20200105125952_add_encrypted_to_messages_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddEncryptedToMessagesTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->boolean('encrypted')->default(false);
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->dropColumn('encrypted');
+        });
+    }
+}